chore: moved ProxyNames to constants
