+++ src/renderer/entities/proxy/index.ts
@@ -1,3 +1,4 @@
 export { proxyUtils } from './lib/utils';
+export { ProxyTypeName } from './lib/constants';
 export { proxyModel } from './model/proxy-model';
 export { ProxyAccount } from './ui/ProxyAccount/ProxyAccount';
