+++ src/renderer/entities/proxy/ui/ProxyAccount/ProxyAccount.tsx
@@ -4,17 +4,7 @@ import { cnTw, toAddress, toShortAddress } from '@shared/lib/utils';
 import { BodyText, HelpText, Identicon, Truncate } from '@shared/ui';
 import { AccountId, ProxyType } from '@shared/core';
 import { useI18n } from '@app/providers';
-
-export const ProxyTypeName: Record<ProxyType, string> = {
-  [ProxyType.ANY]: 'proxy.names.any',
-  [ProxyType.NON_TRANSFER]: 'proxy.names.nonTransfer',
-  [ProxyType.STAKING]: 'proxy.names.staking',
-  [ProxyType.AUCTION]: 'proxy.names.auction',
-  [ProxyType.CANCEL_PROXY]: 'proxy.names.cancelProxy',
-  [ProxyType.GOVERNANCE]: 'proxy.names.governance',
-  [ProxyType.IDENTITY_JUDGEMENT]: 'proxy.names.identityJudgement',
-  [ProxyType.NOMINATION_POOLS]: 'proxy.names.nominationPools',
-};
+import { ProxyTypeName } from '../../lib/constants';
 
 type Props = {
   className?: string;
