+++ src/renderer/features/proxies/lib/proxies-utils.ts
@@ -39,8 +39,8 @@ function getNotification({
       dateCreated: Date.now(),
       proxyType: proxied.proxyType,
       proxyAccountId: proxied.proxyAccountId,
-      proxyWalletName: accountsWalletsMap[proxied.proxyAccountId]?.name,
-      proxyWalletType: accountsWalletsMap[proxied.proxyAccountId]?.type,
+      proxyWalletName: accountsWalletsMap[proxied.proxyAccountId].name,
+      proxyWalletType: accountsWalletsMap[proxied.proxyAccountId].type,
       proxiedAccountId: proxied.accountId,
       proxiedWalletName: proxyUtils.getProxiedName(proxied.accountId, proxied.proxyType, addressPrefix),
       read: false,
